Refactor code
The result property introduced at the end of the performOperation
method is removed for code consistency.

The result property was only being used to hold the result of the
operation before transferring its value to previousOperand for a
subsequent calculation. However, I could just store the result
of the operation as the currentOperand, as if we had just inputted it
for a new operation, since that would be its purpose anyway, if a
user decides to chain an operation off of their previous result.

Refactoring to only depend on the previousOperand and
currentOperand properties makes the code more concise as we don't
have to depend on a third property that gets juggled around anyway.
This also eliminates a redundant switch-like state that the result
property initially acted as.

This change also implemented the following functionality cases to
behave as expected:

Users can perform an operation with the equals button. Users can
then select an operator and perform a chaining calculation using
their previous calculation result as an operand.

Users can chain calulations without pressing the equals button,
evaluating the operation by only selecting a new operator each time.

Users can decide to change their selected operator midway through
a calculation without effects to any other properties.